Remove validation from ArrayDefinition
